Full fix for SwirlModule round trip The first part of the fix is to modify the schema used to read in the pointer table. The SwirlModule modifies the Rom to use a 4 byte pointer instead of a 2 byte pointer, so we need to set the size of the pointer in the schema to 4 so we read the pointers correctly.
The second part of the fix is that we need to read in the offset of the
pointer table from RELOCATED_SWIRL_ANIMATION_POINTER_TABLE_POINTERS,
rather than just reference the location directly like the original Rom
does with SWIRL_ANIMATION_POINTER_TABLE_DEFAULT_OFFSET.

The third part of the fix is to no longer binary OR the read in
addresses with a base address. The 4 byte addresses can reference
anywhere in the Rom.

Lastly, IpsPatch's is_applied() method would always fail because it compared a bytearray
to a list, so to fix, we convert the bytearray to list before
comparison.