Improve UI resize behavior This changes the Notebook at the top of the window to stay the same size as the user resizes the window, while the Text control at the bottom of the window fills the remaining space.
Additionally, the Entry controls in the window will fill horizontally as
the user resizes the window, but Labels and Buttons will stay a fixed
size.

Finally, logic is added to the <<NotebookTabChanged>> event so the
height of the notebook matches the height of the currently selected tab,
rather than always being sized to the height of the tallest tab.