Fix CoilSnake's build (#297)
* Modernize build methods and instructions

So far, this works well enough to install the package "successfully." I
haven't tested whether it actually works or not, or whether the git
version number integration works properly, or whether you can make an
EXE. (I used Python 3.10.6 on Windows 10 for testing so far. I know of
issues with the spec file.)

While I was here, I tried moving everything I could out of setup.py and
into the new file format shared by all build systems. Maybe that was a
bit unnecessary... But I figure pyproject.toml is easier to analyze with
tools, for people who want to scour GitHub and identify security
problems in popular packages and stuff like that.

* Update CI to match the current build process progress

And remove a line of setup.py that I accidentally added while debugging
why some headers couldn't be found in the exhal folder...

* Build EXEs and fix setuptools_scm version junk