CompressedGraphicsModule: fix bad FREE_RANGES end address
FREE_RANGES are inclusive address ranges, and E1F203 is the address of
the spritemap data for the hamburger shop map marker. In rare
circumstances, this could cause the very first byte (Y position) of the
first portion of the sprite to be overwritten with arbitrary data, if
that block of free space was filled all the way to the end.

Probably fixes #316